Refactor modifiers and integrators using common base classes -all float modifiers inherit from EffectModifier_Float, color modifiers from EffectModifier_Color, etc -all float integrators inherit from EffectIntegrator_Float, etc -Refactored Integrator.Update into Update + Apply, where Update is the shared logic between all integrators and Apply is the logic for a specific integrator (taking the output values and applying it to the material, lights, transforms, etc). -Added missing profile markup to the other integrator types
