+++ Source/Waterfall/EffectModifiers/EffectLightFloatModifier.cs
@@ -6,10 +6,12 @@ namespace Waterfall
   /// <summary>
   ///   Material color modifier
   /// </summary>
-  public class EffectLightFloatModifier : EffectModifier
+  public class EffectLightFloatModifier : EffectModifier_Float
   {
+    protected override string ConfigNodeName => WaterfallConstants.LightFloatModifierNodeName;
+
     [Persistent] public string floatName = "";
-    public FloatCurve curve = new();
+    
     private Light[] l;
     public override bool ValidForIntegrator => !string.IsNullOrEmpty(floatName);
 
@@ -20,21 +22,6 @@ namespace Waterfall
 
     public EffectLightFloatModifier(ConfigNode node) : base(node) { }
 
-    public override void Load(ConfigNode node)
-    {
-      base.Load(node);
-      curve.Load(node.GetNode("floatCurve"));
-    }
-
-    public override ConfigNode Save()
-    {
-      var node = base.Save();
-
-      node.name = WaterfallConstants.LightFloatModifierNodeName;
-      node.AddNode(Utils.SerializeFloatCurve("floatCurve", curve));
-      return node;
-    }
-
     public override void Init(WaterfallEffect parentEffect)
     {
       base.Init(parentEffect);
@@ -42,21 +29,6 @@ namespace Waterfall
       for (int i = 0; i < xforms.Count; i++)
       {
         l[i] = xforms[i].GetComponent<Light>();
-      }
-    }
-
-    public void Get(float[] input, float[] output)
-    {
-      if (input.Length > 1)
-      {
-        for (int i = 0; i < l.Length; i++)
-          output[i] = curve.Evaluate(input[i]) + randomValue;
-      }
-      else if (input.Length == 1)
-      {
-        float data = curve.Evaluate(input[0]) + randomValue;
-        for (int i = 0; i < l.Length; i++)
-          output[i] = data;
       }
     }
