+++ Source/Waterfall/Effects/EffectIntegrators/EffectLightColorIntegrator.cs
@@ -4,12 +4,9 @@ using UnityEngine;
 
 namespace Waterfall
 {
-  public class EffectLightColorIntegrator : EffectIntegrator
+  public class EffectLightColorIntegrator : EffectIntegrator_Color
   {
     public string                         colorName;
-    protected readonly Color[] modifierData;
-    protected readonly Color[] initialValues;
-    protected readonly Color[] workingValues;
 
     private readonly Light[]     l;
 
@@ -18,9 +15,6 @@ namespace Waterfall
       // light-color specific
       colorName        = floatMod.colorName;
       l = new Light[xforms.Count];
-      modifierData = new Color[xforms.Count];
-      initialValues = new Color[xforms.Count];
-      workingValues = new Color[xforms.Count];
 
       for (int i = 0; i < xforms.Count; i++)
       {
@@ -29,22 +23,8 @@ namespace Waterfall
       }
     }
 
-    public override void Update()
+    protected override void Apply()
     {
-      if (handledModifiers.Count == 0)
-        return;
-      Array.Copy(initialValues, workingValues, l.Length);
-
-      foreach (var mod in handledModifiers)
-      {
-        if (mod.Controller != null)
-        {
-          float[] controllerData = mod.Controller.Get();
-          ((EffectLightColorModifier)mod).Get(controllerData, modifierData);
-          Integrate(mod.effectMode, workingValues, modifierData);
-        }
-      }
-
       for (int i = 0; i < l.Length; i++)
         l[i].color = workingValues[i];
     }
