+++ Source/Waterfall/Effects/EffectIntegrators/EffectLightFloatIntegrator.cs
@@ -5,12 +5,9 @@ using UnityEngine;
 
 namespace Waterfall
 {
-  public class EffectLightFloatIntegrator : EffectIntegrator
+  public class EffectLightFloatIntegrator : EffectIntegrator_Float
   {
     public string                         floatName;
-    protected readonly float[] modifierData;
-    protected readonly float[] initialValues;
-    protected readonly float[] workingValues;
 
     private readonly Light[]     l;
 
@@ -23,9 +20,6 @@ namespace Waterfall
       testIntensity = WaterfallConstants.ShaderPropertyHideFloatNames.Contains(floatName);
 
       l = new Light[xforms.Count];
-      modifierData = new float[xforms.Count];
-      initialValues = new float[xforms.Count];
-      workingValues = new float[xforms.Count];
 
       for (int i = 0; i < xforms.Count; i++)
       {
@@ -37,23 +31,8 @@ namespace Waterfall
       }
     }
 
-    public override void Update()
+    protected override void Apply()
     {
-      if (!Settings.EnableLights || handledModifiers.Count == 0)
-        return;
-
-      Array.Copy(initialValues, workingValues, l.Length);
-
-      foreach (var mod in handledModifiers)
-      {
-        if (mod.Controller != null)
-        {
-          float[] controllerData = mod.Controller.Get();
-          ((EffectLightFloatModifier)mod).Get(controllerData, modifierData);
-          Integrate(mod.effectMode, workingValues, modifierData);
-        }
-      }
-
       float lightBaseScale = parentEffect.TemplateScaleOffset.x;
       for (int i = 0; i < l.Length; i++)
       {
