+++ Source/Waterfall/Effects/EffectIntegrators/EffectScaleIntegrator.cs
@@ -4,38 +4,16 @@ using UnityEngine;
 
 namespace Waterfall
 {
-  public class EffectScaleIntegrator : EffectIntegrator
-  {
-    protected readonly Vector3[] modifierData;
-    protected readonly Vector3[] initialValues;
-    protected readonly Vector3[] workingValues;
-    
+  public class EffectScaleIntegrator : EffectIntegrator_Vector3
+  { 
     public EffectScaleIntegrator(WaterfallEffect effect, EffectScaleModifier mod) : base(effect, mod)
     {
-      modifierData = new Vector3[xforms.Count];
-      initialValues = new Vector3[xforms.Count];
-      workingValues = new Vector3[xforms.Count];
-
       for(int i = 0; i < xforms.Count; i++)
         initialValues[i] = xforms[i].localScale;
     }
 
-    public override void Update()
+    protected override void Apply()
     {
-      if (handledModifiers.Count == 0)
-        return;
-      Array.Copy(initialValues, workingValues, initialValues.Length);
-
-      foreach (var mod in handledModifiers)
-      {
-        if (mod.Controller != null)
-        {
-          float[] controllerData = mod.Controller.Get();
-          ((EffectScaleModifier)mod).Get(controllerData, modifierData);
-          Integrate(mod.effectMode, workingValues, modifierData);
-        }
-      }
-
       for (int i = 0; i < xforms.Count; i++)
         xforms[i].localScale = workingValues[i];
     }
